Move groups.GroupsPrefTab to gui.preftabs
